Custom Graphics branch had been merged.
git-svn-id: http://chianti.ucsd.edu/svn@20380 0ecc0d97-ab19-0410-9704-bfe1a75892f5

+++ cytoscape/trunk/src/cytoscape/visual/ui/editors/continuous/C2DMappingEditor.java
@@ -318,7 +318,7 @@ public class C2DMappingEditor extends ContinuousMappingEditorPanel {
 		for (ContinuousMappingPoint point : allPoints) {
 			bound = point.getRange();
 
-			fraction = ((Number) ((point.getValue() - minValue) / actualRange)).floatValue() * 100;
+			fraction = ((Number) ((point.getValue().doubleValue() - minValue) / actualRange)).floatValue() * 100;
 			slider.getModel().addThumb(fraction, bound.equalValue);
 		}
